requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
pytest==8.2.0
httpx==0.27.0
alembic
psycopg2-binary